new search results common components
